requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
PyJWT>=2.8
psycopg2>=2.9
python-dotenv>=1.0
